+++ src/index.css
@@ -95,7 +95,7 @@ button:disabled {
 }
 
 .card:hover {
-  box-shadow: 0 8px 16px 0 rgb(255, 231, 143);
+  box-shadow: 0 8px 16px 8px rgb(255, 231, 143);
 }
 
 .cardButtonGroup {
